setup axios get for questions data
