+++ client/src/components/App.jsx
@@ -1,6 +1,5 @@
 // imports the react module installed in node_modules
 import React, { useState } from 'react';
-import axios from 'axios';
 // import Overview from './overview/Overview.jsx';
 // import Related from './related/Related.jsx';
 import Question from './question-answer/Question.jsx';
