+++ server/index.js
@@ -19,7 +19,6 @@ const sessionConfig = {
 };
 app.use(express.json());
 app.use(serveStatic);
-
 app.use(session(sessionConfig));
 
 // PRODUCT API Calls
@@ -116,32 +115,31 @@ app.get('/reviews/meta/:product_id', (req, res) => {
 // Q&A API Calls
 app.get('/qa/questions/:product_id', (req, res) => {
   axios({
-    url: `${API}/qa/questions?product_id=${req.params.product_id}&page=1&count=5`,
+    url: `${API}/qa/questions?product_id=${req.params.product_id}`,
     method: 'GET',
-    headers: { Authorization: APIInfo.token }
+    headers: { Authorization: APIInfo.token },
   })
     .then(response => {
       res.status(200).send(response.data);
     })
     .catch(error => {
       res.status(500).send(error);
-    })
-})
-
+    });
+});
 
 app.get('/qa/questions/:question_id/answers', (req, res) => {
   axios({
-    url: `${API}/qa/questions/${req.params.question_id}/answers&page=1&count=5`,
+    url: `${API}/qa/questions/${req.params.question_id}`,
     method: 'GET',
-    headers: { Authorization: APIInfo.token }
+    headers: { Authorization: APIInfo.token },
   })
-    .then(response => {
+    .then((response) => {
       res.status(200).send(response.data);
     })
-    .catch(error => {
+    .catch((error) => {
       res.status(500).send(error);
-    })
-})
+    });
+});
 
 // ***********************************************************************
 app.listen(PORT, () => {
